rewrite for chef 11
