+++ .kitchen.yml
@@ -2,7 +2,7 @@
 driver:
   name: docker
 driver_config:
-  require_chef_omnibus: true
+  require_chef_omnibus: 11.16.4
   use_sudo: false
 
 provisioner:
