+++ metadata.rb
@@ -5,5 +5,3 @@ license 'Apache 2.0'
 description 'Installs/Configures spark'
 long_description 'Installs/Configures spark'
 version '0.1.0'
-
-depends 'compat_resource', '>= 12.10.4'
